updated sample
