+++ samples/client/petstore/android-java/src/main/java/io/swagger/client/JsonUtil.java
@@ -1,4 +1,4 @@
-package com.wordnik.client;
+package io.swagger.client;
 
 import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.databind.*;
